Fix VS2026/ClangCL builds and stabilize CTest on Windows.
Add CMake presets, smoke tests, and an example ABI shim so IntelDFPExample and IntelDFPTest pass under MSVC and ClangCL while suppressing remaining conversion warnings.

+++ LIBRARY/src/bid128_string.c
@@ -216,7 +216,7 @@ bid128_to_string (char *str, BID_UINT128 x
 	__L1_Split_MiDi_6_Lead (HI_18Dig, ptr);
 	__L1_Split_MiDi_6 (LO_18Dig, ptr);
       }
-      len = ptr - MiDi;
+      len = (int)(ptr - MiDi);
       c_ptr_start = &(str[k]);
       c_ptr = c_ptr_start;
 
@@ -225,7 +225,7 @@ bid128_to_string (char *str, BID_UINT128 x
       for (k_lcv = 1; k_lcv < len; k_lcv++) {
 	__L0_MiDi2Str (MiDi[k_lcv], c_ptr);
       }
-      k = k + (c_ptr - c_ptr_start);
+      k = k + (unsigned int)(c_ptr - c_ptr_start);
     }
 
     // print E and sign of exponent
