+++ LIBRARY/src/bid64_string.c
@@ -188,7 +188,7 @@ bid64_to_string (char *ps, BID_UINT64 x
 
       ptr = MiDi;
       __L1_Split_MiDi_6_Lead (LO_18Dig, ptr);
-      len = ptr - MiDi;
+      len = (int)(ptr - MiDi);
       c_ptr_start = &(ps[istart]);
       c_ptr = c_ptr_start;
 
@@ -197,7 +197,7 @@ bid64_to_string (char *ps, BID_UINT64 x
       for (k_lcv = 1; k_lcv < len; k_lcv++) {
         __L0_MiDi2Str (MiDi[k_lcv], c_ptr);
       }
-      istart = istart + (c_ptr - c_ptr_start);
+      istart = istart + (int)(c_ptr - c_ptr_start);
     }
 
     ps[istart++] = 'E';
